@@ -2,12 +2,12 @@ package com.example.justin.resume_internal
 
 import android.graphics.Typeface
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.view.animation.Animation
 import android.view.animation.AnimationUtils
+import androidx.fragment.app.Fragment
 
 import kotlinx.android.synthetic.main.background_scroll.*
 
